Move component defaults into components.
