--- src/libs/context/config/README.md
@@ -1,7 +1,0 @@
-# Context - Config
-
-Look for User Configuration, and if it exists load it so can be used by the custom `Context` library.
-
-## description
-
-- [ ] TODO: Add detailed description
