--- src/libs/context/defaults/README.md
@@ -1,5 +1,0 @@
-# Context - Defaults
-
-## description
-
-A centralized location for default values.
