--- src/libs/context/types/READEME.md
@@ -1,3 +1,0 @@
-# Types
-
-Universal and shared types for project are stored here to allow for easy access and reuse.
